fix: move struct to its own file and private the inner regis
